consistent naming for upgradehelper.py
Rename upgradehelper.py into upgrade-helper.py since the conf file and the git
repo are named with a dash to separate "upgrade" and "helper".

The documentation in the README and upgradehelper.py were also already using
upgrade-helper.py instead of upgradehelper.py.

For all those reasons, let's be consistent and rename the python script.

Signed-off-by: Quentin Schulz <foss@0leil.net>