upgrade-helper: rewrite the code to update groups of recipes in lockstep
This support the changes in oe-core that group recipes if they share
common includes, and allow lockstep updates of items like
cmake and cmake-native, that previously would reliably fail when
updated separately.

The patch is invasive: the code throughout assumes that each
update transaction is for a single recipe, and adjusting that
to split the data into recipe-specific and update-specific parts
has to happen all over the place.

Signed-off-by: Alexander Kanavin <alex@linutronix.de>